Refractoring and made diffrent file for calling itune api
