+++ app/services/ituneApi.js
@@ -1,0 +1,5 @@
+import { generateApiClient } from '@app/utils/apiUtils';
+
+const ituneApi = generateApiClient('itune');
+
+export const getItunes = (ituneName) => ituneApi.get(`/search?term=${ituneName}`);
