+++ app/services/tests/repoApi.test.js
@@ -1,6 +1,6 @@
 import MockAdapter from 'axios-mock-adapter';
 import { getApiClient } from '@utils/apiUtils';
-import { getItunes, getRepos } from '../repoApi';
+import { getRepos } from '../repoApi';
 
 describe('RepoApi tests', () => {
   const repositoryName = 'mac';
@@ -14,20 +14,6 @@ describe('RepoApi tests', () => {
     ];
     mock.onGet(`/search/repositories?q=${repositoryName}`).reply(200, data);
     const res = await getRepos(repositoryName);
-    expect(res.data).toEqual(data);
-  });
-
-  const ituneName = 'test';
-  it('should make the api call to "/search?term=', async () => {
-    const mock = new MockAdapter(getApiClient('itune').axiosInstance);
-    const data = [
-      {
-        totalCount: 1,
-        items: [{ ituneName }]
-      }
-    ];
-    mock.onGet(`/search?term=${ituneName}`).reply(200, data);
-    const res = await getItunes(ituneName);
     expect(res.data).toEqual(data);
   });
 });
